+++ app/[username]/[slug]/page.js
@@ -1,0 +1,6 @@
+import PostComponent from "../../../src/pages/[username]/PostComponent";
+async function Post() {
+  // Add API logic here and pass results as props to PostComponent
+  return <PostComponent />;
+}
+export default Post;
