+++ app/[username]/page.js
@@ -4,27 +4,27 @@ import Head from 'next/head'
 import { css } from '@emotion/react'
 import { htmlToText } from 'html-to-text'
 
-import { truncate } from '../../lib/utils'
-import { getUserByName } from '../../lib/db'
+import { truncate } from '../../src/lib/utils'
+import { getUserByName } from '../../src/lib/db'
 
-import meta from '../../components/meta'
-import Container from '../../components/container'
+import meta from '../../src/components/meta'
+import Container from '../../src/components/container'
 
 export default function Profile({ user }) {
   return (
     <Container maxWidth="640px">
-      <Head>
-        {meta({
-          title: `${user.displayName} (@${user.name}) / The Abyss`,
-          description: user.about,
-          url: `/${user.name}`,
-          image: user.photo,
-        })}
-        <link
-          href="https://fonts.googleapis.com/css2?family=Newsreader:ital,wght@0,400;0,600;1,400;1,600&display=swap"
-          rel="stylesheet"
-        />
-      </Head>
+      <Helmet>
+  {meta({
+    title: `${user.displayName} (@${user.name}) / The Abyss`,
+    description: user.about,
+    url: `/${user.name}`,
+    image: user.photo,
+  })}
+  <link
+    href="https://fonts.googleapis.com/css2?family=Newsreader:ital,wght@0,400;0,600;1,400;1,600&display=swap"
+    rel="stylesheet"
+  />
+</Helmet>
 
       <img
         src={user.photo}
@@ -102,42 +102,41 @@ export default function Profile({ user }) {
             </p>
 
             <Link href={`/${user.name}/${post.slug}`}>
-              <a>
-                <h3
-                  css={css`
-                    font-size: 1rem;
-                    font-weight: 400;
-                    margin-bottom: 0.6rem;
-                  `}
-                >
-                  {post.title ? htmlToText(post.title) : 'Untitled'}
-                </h3>
 
-                <p
-                  css={css`
-                    max-width: 25rem;
-                    color: var(--grey-4);
-                    font-family: 'Newsreader', serif;
-                    line-height: 1.5em;
-                  `}
-                >
-                  {post.excerpt
-                    ? htmlToText(post.excerpt)
-                    : truncate(htmlToText(post.content), 25)}
-                </p>
-              </a>
+              <h3
+                css={css`
+                  font-size: 1rem;
+                  font-weight: 400;
+                  margin-bottom: 0.6rem;
+                `}
+              >
+                {post.title ? htmlToText(post.title) : 'Untitled'}
+              </h3>
+              <p
+                css={css`
+                  max-width: 25rem;
+                  color: var(--grey-4);
+                  font-family: 'Newsreader', serif;
+                  line-height: 1.5em;
+                `}
+              >
+                {post.excerpt
+                  ? htmlToText(post.excerpt)
+                  : truncate(htmlToText(post.content), 25)}
+              </p>
+
             </Link>
           </li>
         ))}
       </ul>
     </Container>
-  )
+  );
 }
 
-export async function getStaticPaths() {
+export async function generateStaticParams() {
   return {
+    dynamicParams: 'blocking',
     paths: [],
-    fallback: 'blocking',
   }
 }
