+++ app/dashboard/[pid]/page.js
@@ -1,0 +1,6 @@
+import PostEditorComponent from "../../../src/pages/dashboard/PostEditorComponent";
+async function PostEditor() {
+  // Add API logic here and pass results as props to PostEditorComponent
+  return <PostEditorComponent />;
+}
+export default PostEditor;
