+++ app/dashboard/list/page.js
@@ -1,0 +1,6 @@
+import ReadingListComponent from "../../../src/pages/dashboard/ReadingListComponent";
+async function ReadingList() {
+  // Add API logic here and pass results as props to ReadingListComponent
+  return <ReadingListComponent />;
+}
+export default ReadingList;
