+++ app/dashboard/page.js
@@ -1,0 +1,6 @@
+import DashboardComponent from "../../src/pages/dashboard/DashboardComponent";
+async function Dashboard() {
+  // Add API logic here and pass results as props to DashboardComponent
+  return <DashboardComponent />;
+}
+export default Dashboard;
