+++ app/page.js
@@ -1,0 +1,6 @@
+import HomeComponent from "../src/pages/HomeComponent";
+async function Home() {
+  // Add API logic here and pass results as props to HomeComponent
+  return <HomeComponent />;
+}
+export default Home;
