+++ package.json
@@ -7,28 +7,23 @@
     "build": "next build",
     "start": "npm run start"
   },
-  
-  "engines" : { 
-    "node" : "16.20.2"
-  },
-  
+  "engines": {},
   "config": {
     "legacy-peer-deps": true
   },
-  
   "dependencies": {
     "@emotion/react": "^11.4.0",
     "@radix-ui/react-dialog": "^0.0.19",
     "@radix-ui/react-icons": "^1.0.3",
     "@radix-ui/react-id": "^0.0.6",
     "@react-firebase/auth": "^0.2.10",
+    "@tiptap/extension-highlight": "^2.1.8",
     "@tiptap/extension-image": "^2.0.0",
     "@tiptap/extension-link": "^2.0.0",
     "@tiptap/extension-placeholder": "^2.0.0",
+    "@tiptap/pm": "^2.0.0",
     "@tiptap/react": "^2.0.0",
     "@tiptap/starter-kit": "^2.0.0",
-    "@tiptap/pm": "^2.0.0",
-    "@tiptap/extension-highlight":"^2.1.8",
     "firebase": "^8.7.1",
     "html-to-text": "^8.0.0",
     "modern-normalize": "^1.1.0",
